backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==20.1.0
SQLAlchemy==2.0.15
pytest==7.3.1
sqlite3